Add support for prebuilt deb packages
This commit switches the default build to deb packages
to allow use of widely known apt/dpkg tools.

It also adds to the image the tools required to install extra
packages that are pre-built by Synaptics and available on a
public server.

It also removes the db recipe as this is not required in
the base configuration and this recipe is not compatible with
apt/dpkg.

Change-Id: I1427f58f641f46ae7d97fa5dcff92af7d8753227
Reviewed-on: https://sc-debu-git.synaptics.com/gerrit/227099
Reviewed-by: Lorenzo Keller <Lorenzo.Keller@synaptics.com>
Reviewed-by: Dharini Subramaniam <Dharini.Subramaniam@synaptics.com>
Reviewed-by: Tao Yu <Tao.Yu@synaptics.com>